# docs/swagger.yaml
basePath: /
definitions:
...
  monitors.CreateMonitorRequest:
    properties:
      accepted_status_codes:
        items:
          type: string
        type: array
      auth_password:
        type: string
      auth_username:
        type: string
      body:
        type: string
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      expected_value:
        type: string
      filters_contains:
        type: string
      filters_not_contains:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      ignore_tls_errors:
        type: boolean
      interval:
        minimum: 1
        type: integer
      json_path:
        type: string
      max_redirects:
        minimum: 0
        type: integer
      method:
        minLength: 1
        type: string
      name:
        minLength: 1
        type: string
      push_token:
        type: string
      retry:
        minimum: 0
        type: integer
      retry_after:
        minimum: 0
        type: integer
      timeout:
        minimum: 1
        type: integer
      type:
        enum:
        - http
        - keyword
        - json
        - ping
        - tcp
        - smtp
        - dns
        - push
        - grpc
        type: string
      url:
        minLength: 1
        type: string
    required:
    - name
    - type
    - url
    type: object
  monitors.MonitorCheckResponse:
    properties:
      checked_at:
        type: string
      created_at:
        type: string
      duration_seconds:
        type: integer
      error:
        type: string
      id:
        type: integer
      important:
        type: boolean
      message:
        type: string
      monitor_id:
        type: integer
      response_time_ms:
        type: integer
      retry_count:
        type: integer
      status:
        type: string
      status_code:
        type: integer
      updated_at:
        type: string
    type: object
  monitors.MonitorChecksResponse:
    properties:
      checks:
        items:
          $ref: '#/definitions/monitors.MonitorCheckResponse'
        type: array
    type: object
  monitors.MonitorListResponse:
    properties:
      monitors:
        items:
          $ref: '#/definitions/monitors.MonitorResponse'
        type: array
    type: object
  monitors.MonitorResponse:
    properties:
      accepted_status_codes:
        items:
          type: string
        type: array
      auth_username:
        type: string
      body:
        type: string
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      description:
        type: string
      expected_value:
        type: string
      filters_contains:
        type: string
      filters_not_contains:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      id:
        type: integer
      ignore_tls_errors:
        type: boolean
      interval:
        type: integer
      is_active:
        type: boolean
      json_path:
        type: string
      max_redirects:
        type: integer
      method:
        type: string
      name:
        type: string
      push_token:
        type: string
      retry:
        type: integer
      retry_after:
        type: integer
      status:
        type: string
      timeout:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: integer
    type: object
...
  monitors.UpdateMonitorRequest:
    properties:
      accepted_status_codes:
        items:
          type: string
        type: array
      auth_password:
        type: string
      auth_username:
        type: string
      body:
        type: string
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      expected_value:
        type: string
      filters_contains:
        type: string
      filters_not_contains:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      ignore_tls_errors:
        type: boolean
      interval:
        minimum: 1
        type: integer
      json_path:
        type: string
      max_redirects:
        minimum: 0
        type: integer
      method:
        minLength: 1
        type: string
      name:
        minLength: 1
        type: string
      push_token:
        type: string
      retry:
        minimum: 0
        type: integer
      retry_after:
        minimum: 0
        type: integer
      timeout:
        minimum: 1
        type: integer
      type:
        enum:
        - http
        - keyword
        - json
        - ping
        - tcp
        - smtp
        - dns
        - push
        - grpc
        type: string
      url:
        minLength: 1
        type: string
    type: object
  routes.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  routes.SuccessResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
host: 127.0.0.1:3031
info:
  contact: {}
  description: PhoenixUptime Backend API
  title: PhoenixUptime API
  version: "1.0"
paths:
...
  /monitors:
    get:
      consumes:
      - application/json
      description: Returns monitors owned by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: monitor list
          schema:
            $ref: '#/definitions/monitors.MonitorListResponse'
        "401":
          description: unauthorized - invalid or missing API key
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - ApiKeyQuery: []
      summary: List Monitors
      tags:
      - Monitors
    post:
      consumes:
      - application/json
      description: Creates a monitor owned by the authenticated user.
      parameters:
      - description: monitor payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/monitors.CreateMonitorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created monitor
          schema:
            $ref: '#/definitions/monitors.MonitorResponse'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: unauthorized - invalid or missing API key
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - ApiKeyQuery: []
      summary: Create Monitor
      tags:
      - Monitors
  /monitors/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes one monitor owned by the authenticated user.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: monitor deleted
          schema:
            $ref: '#/definitions/routes.SuccessResponse'
        "400":
          description: invalid monitor id
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: unauthorized - invalid or missing API key
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: monitor not found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - ApiKeyQuery: []
      summary: Delete Monitor
      tags:
      - Monitors
    get:
      consumes:
      - application/json
      description: Returns one monitor owned by the authenticated user.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: monitor
          schema:
            $ref: '#/definitions/monitors.MonitorResponse'
        "400":
          description: invalid monitor id
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: unauthorized - invalid or missing API key
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: monitor not found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - ApiKeyQuery: []
      summary: Get Monitor
      tags:
      - Monitors
    patch:
      consumes:
      - application/json
      description: Updates one monitor owned by the authenticated user.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      - description: monitor payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/monitors.UpdateMonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: updated monitor
          schema:
            $ref: '#/definitions/monitors.MonitorResponse'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: unauthorized - invalid or missing API key
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: monitor not found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - ApiKeyQuery: []
      summary: Update Monitor
      tags:
      - Monitors
  /monitors/{id}/checks:
    get:
      consumes:
      - application/json
      description: Returns recent check results for one monitor owned by the authenticated
        user.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Maximum checks to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: monitor checks
          schema:
            $ref: '#/definitions/monitors.MonitorChecksResponse'
        "400":
          description: invalid input
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: unauthorized - invalid or missing API key
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: monitor not found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyHeader: []
      - ApiKeyQuery: []
      summary: List Monitor Checks
      tags:
      - Monitors
...
securityDefinitions:
  ApiKeyHeader:
    in: header
    name: x-api-key
    type: apiKey
  ApiKeyQuery:
    in: query
    name: api_key
    type: apiKey
swagger: "2.0"